% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JGCRIutils.R
\docType{package}
\name{JGCRIutils}
\alias{JGCRIutils}
\alias{JGCRIutils-package}
\title{JGCRI utilities}
\description{
A set of utilities supporting logging, working with data, data visualization,
and reproducibility at the Joint Global Change Research Institute.
}

