% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logging.R
\name{openlog}
\alias{openlog}
\title{Open a new logfile}
\usage{
openlog(scriptname, loglevel = -Inf, logfile = NULL, append = FALSE,
  sink = TRUE)
}
\arguments{
\item{scriptname}{Name of script (and thus logfile)}

\item{loglevel}{Minimum priority level (numeric, optional)}

\item{logfile}{Override default logfile (character, optional)}

\item{append}{Append to logfile? (logical, optional)}

\item{sink}{Sink to logfile? (logical, optional)}
}
\value{
Invisible success (TRUE) or failure (FALSE)
}
\description{
Open a new logfile
}
\details{
Open a new logfile. Note that if \code{sink} is TRUE, all
screen output will be captured (via \code{\link{base::sink}}).
Re-opening a logfile will erase the previous output unless \code{append}
is TRUE. Finally, messages will only appear in the logfile if their
\code{level} exceeds \code{loglevel}.
}
\seealso{
\code{\link{printlog}} \code{\link{closelog}}
}

