% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logging.R
\name{printlog}
\alias{printlog}
\title{Time-stamped output function}
\usage{
printlog(msg = "", ..., level = 0, ts = TRUE, cr = TRUE)
}
\arguments{
\item{msg}{One or more messages to log (optional)}

\item{level}{Priority level (numeric, optional)}

\item{ts}{Print preceding timestamp? (logical, optional)}

\item{cr}{Print trailing newline? (logical, optional)}
}
\value{
Invisible success (TRUE) or failure (FALSE)
}
\description{
Time-stamped output function
}
\details{
Logs a message, which may consist of one or more printable objects
}
\seealso{
\code{\link{openlog}} \code{\link{closelog}}
}

